v0/requirements_build.txt
emoji
numpy
scikit-learn
umap-learn
plotly
pandas
gdown
kaleido
